// src/web/shared/proto/jaiabot/messages/jaia_dccl.ts
export interface Command {
    bot_id?: number | undefined;
    time?: string | undefined;
    /**
     * Special case: not sent in message - written upon receipt
     * link: which link this message traversed from hub to bot
     */
    link?: Link | undefined;
    from_hub_id?: number | undefined;
    type?: Command_CommandType | undefined;
    /** required for type == MISSION_PLAN */
    plan?: MissionPlan | undefined;
    /** required for type == REMOTE_CONTROL_SETPOINT */
    rc?: RemoteControl | undefined;
    /** required for type == REMOTE_CONTROL_TASK */
    rc_task?: MissionTask | undefined;
}

export enum Command_CommandType {
    /** MISSION_PLAN - pre mission */
    MISSION_PLAN = "MISSION_PLAN",
    ACTIVATE = "ACTIVATE",
    START_MISSION = "START_MISSION",
    MISSION_PLAN_FRAGMENT = "MISSION_PLAN_FRAGMENT",
    /** NEXT_TASK - during any mission */
    NEXT_TASK = "NEXT_TASK",
    RETURN_TO_HOME = "RETURN_TO_HOME",
    STOP = "STOP",
    PAUSE = "PAUSE",
    RESUME = "RESUME",
    /** REMOTE_CONTROL_SETPOINT - during remote control mission */
    REMOTE_CONTROL_SETPOINT = "REMOTE_CONTROL_SETPOINT",
    REMOTE_CONTROL_TASK = "REMOTE_CONTROL_TASK",
    REMOTE_CONTROL_RESUME_MOVEMENT = "REMOTE_CONTROL_RESUME_MOVEMENT",
    /** RECOVERED - post mission */
    RECOVERED = "RECOVERED",
    SHUTDOWN = "SHUTDOWN",
    RETRY_DATA_OFFLOAD = "RETRY_DATA_OFFLOAD",
    DATA_OFFLOAD_COMPLETE = "DATA_OFFLOAD_COMPLETE",
    DATA_OFFLOAD_FAILED = "DATA_OFFLOAD_FAILED",
    /** RESTART_ALL_SERVICES - debugging low level commands */
    RESTART_ALL_SERVICES = "RESTART_ALL_SERVICES",
    REBOOT_COMPUTER = "REBOOT_COMPUTER",
    SHUTDOWN_COMPUTER = "SHUTDOWN_COMPUTER",
}

export interface CommandForHub {
    hub_id?: number | undefined;
    time?: string | undefined;
    type?: CommandForHub_HubCommandType | undefined;
    scan_for_bot_id?: number | undefined;
    hub_location?: GeographicCoordinate | undefined;
}

export enum CommandForHub_HubCommandType {
    SCAN_FOR_BOTS = "SCAN_FOR_BOTS",
    CTD_DATA_OFFLOAD = "CTD_DATA_OFFLOAD",
    /** RESTART_ALL_SERVICES - debugging low level commands */
    RESTART_ALL_SERVICES = "RESTART_ALL_SERVICES",
    REBOOT_COMPUTER = "REBOOT_COMPUTER",
    SHUTDOWN_COMPUTER = "SHUTDOWN_COMPUTER",
    /** SET_HUB_LOCATION - simulator commands */
    SET_HUB_LOCATION = "SET_HUB_LOCATION",
}